requirements.txt
fastapi
uvicorn
pydantic
pyjwt
authlib
pymongo
psycopg2-binary
pydub
requests
python-dotenv
